refactor: 데이터베이스 패키지 분리
