+++ app/src/main/java/woowacourse/shopping/data/db/CartProductEntity.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data
+package woowacourse.shopping.data.db
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
